Refactor common-lib namespace from address to common
