+++ session/src/main/java/uk/gov/di/ipv/cri/address/api/service/SessionRequestService.java
@@ -6,11 +6,11 @@ import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import uk.gov.di.ipv.cri.address.api.domain.RawSessionRequest;
-import uk.gov.di.ipv.cri.address.library.domain.SessionRequest;
-import uk.gov.di.ipv.cri.address.library.exception.ClientConfigurationException;
-import uk.gov.di.ipv.cri.address.library.exception.SessionValidationException;
-import uk.gov.di.ipv.cri.address.library.service.ConfigurationService;
-import uk.gov.di.ipv.cri.address.library.service.JWTVerifier;
+import uk.gov.di.ipv.cri.common.library.domain.SessionRequest;
+import uk.gov.di.ipv.cri.common.library.exception.ClientConfigurationException;
+import uk.gov.di.ipv.cri.common.library.exception.SessionValidationException;
+import uk.gov.di.ipv.cri.common.library.service.ConfigurationService;
+import uk.gov.di.ipv.cri.common.library.service.JWTVerifier;
 
 import java.net.URI;
 import java.text.ParseException;
